+++ CHANGELOG.md
@@ -26,6 +26,13 @@
   development of standard libraries.
 - Create dedicated `RegionValidationError` and `RegionValidationResult`.
 
+## 0.9.4 (2020-07-16)
+
+**cosmwasm-vm**
+
+- Add `Instance::create_gas_report` returning a gas report including the
+  original limit, the remaining gas and the internally/externally used gas.
+
 ## 0.9.3 (2020-07-08)
 
 **cosmwasm-storage**